+++ src/api/routes.py
@@ -8,7 +8,7 @@ from flask_cors import CORS
 from flask_jwt_extended import create_access_token
 from flask_jwt_extended import get_jwt_identity
 from flask_jwt_extended import jwt_required
-
+from flask_bcrypt import Bcrypt
 
 api = Blueprint('api', __name__)
 
@@ -26,3 +26,30 @@ def handle_hello():
     return jsonify(response_body), 200
 
 
+# Sign Up o Registro
+
+@api.route('/signup', methods=["POST"])
+def signup():
+    body = request.get_json(silent=True)
+    
+    if body is None:
+        return jsonify ({'msg':'Los campos de usuario y contraseña están vacíos'}), 400
+    if 'username' not in body:
+        return jsonify ({'msg':'Debe crear un usuario para continuar.'}), 400
+    if 'email' not in body:
+        return jsonify ({'msg':'No puede continuar sin un correo electrónico.'}), 400
+    if 'password' not in body:
+        return jsonify ({'msg':'No puede continuar sin su contraseña de seguridad.'}), 400
+    
+    password_hash= current_app.bcrypt.generate_password_hash(body["password"]).decode("utf-8")
+
+    new_user = User(
+        username = body["username"],
+        role = body["role"],
+        email = body["email"],
+        password = password_hash,
+        is_active = True,
+    )
+    db.session.add(new_user)
+    db.session.commit()
+    return jsonify ({'msg':'Usuario Creado .'}), 200
