+++ src/front/js/pages/signup.js
@@ -2,20 +2,20 @@ import React, { useContext, useState } from "react";
 import { Context } from "../store/appContext";
 import { useNavigate } from "react-router-dom";
 
-
 export const Signup = () => {
   const { actions } = useContext(Context);
   const navigate = useNavigate();
   const [username, setUsername] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [role, setRole] = useState("");
 
   const register = async (e) => {
     e.preventDefault();
-    if (username == "" && email == "" && password == "") {
+    if (username == "" && email == "" && password == "" &&role == "") {
       alert("Hay campos vacios ");
     } else {
-      const result = await actions.register(username, email, password);
+      const result = await actions.register(username, email, password, role);
       if (result) {
         navigate("/");
       }
@@ -72,6 +72,21 @@ export const Signup = () => {
                 className="form-control"
                 id="exampleInputPassword1"
               />
+            </div>
+            <div className="col">
+            <label htmlFor="exampleInputPassword1" className="form-label">
+                Role
+              </label>
+              <select
+                className="form-select"
+                aria-label="Default select example"
+                value={role}
+                onChange={(e) => setRole(e.target.value)}
+              >
+                <option selected="">select Role </option>
+                <option value="Client">Client</option>
+                <option value="Provider">Provider</option>
+              </select>
             </div>
           </div>
           <div className="d-flex justify-content-center">
